requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
django-phonenumber-field==8.0.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
phonenumbers==9.0.1
pillow
PyJWT==2.9.0
pycparser==2.22
sqlparse==0.5.3
